Remove mocking from tests
