+++ src/tests/banners/bannerSelection.test.ts
@@ -1,17 +1,18 @@
-import { selectBannerTest, _ } from './bannerSelection';
-import { cacheAsync as _cacheAsync } from '../../lib/cache';
+import { selectBannerTest } from './bannerSelection';
+import { getTests } from './bannerTests';
+import { BannerDeployCaches, ReaderRevenueRegion } from './bannerDeployCache';
 
-const cacheAsync = _cacheAsync;
-
-jest.mock('../../lib/cache', () => ({
-    cacheAsync: jest.fn(),
-}));
+const getCache = (date: string, region: ReaderRevenueRegion) =>
+    ({
+        subscriptions: {
+            [region]: () => Promise.resolve(new Date(date)),
+        },
+        contributions: {
+            [region]: () => Promise.resolve(new Date(date)),
+        },
+    } as BannerDeployCaches);
 
 describe('selectBannerTest', () => {
-    afterEach(() => {
-        jest.resetAllMocks();
-    });
-
     const firstDate = 'Mon Jun 06 2020 19:20:10 GMT+0100';
     const secondDate = 'Mon Jul 06 2020 19:20:10 GMT+0100';
 
@@ -37,38 +38,23 @@ describe('selectBannerTest', () => {
         };
 
         it('returns banner if it has never been dismissed', () => {
-            (cacheAsync as jest.Mock).mockReturnValue([
-                null,
-                (): Promise<Date> => Promise.resolve(new Date(secondDate)),
-            ]);
+            const cache = getCache(secondDate, 'united-states');
 
-            _.resetCache('subscriptions', 'united-states');
-
-            return selectBannerTest(targeting, tracking, '').then(result => {
+            return selectBannerTest(targeting, tracking, '', getTests, cache).then(result => {
                 expect(result && result.test.name).toBe('DigitalSubscriptionsBanner');
             });
         });
 
         it('returns banner if has been redeployed', () => {
-            (cacheAsync as jest.Mock).mockReturnValue([
-                null,
-                (): Promise<Date> => Promise.resolve(new Date(firstDate)),
-            ]);
+            const cache = getCache(firstDate, 'united-states');
 
-            _.resetCache('subscriptions', 'united-states');
-
-            return selectBannerTest(targeting, tracking, '').then(result => {
+            return selectBannerTest(targeting, tracking, '', getTests, cache).then(result => {
                 expect(result && result.test.name).toBe('DigitalSubscriptionsBanner');
             });
         });
 
         it('returns null if there are insufficient page views', () => {
-            (cacheAsync as jest.Mock).mockReturnValue([
-                null,
-                (): Promise<Date> => Promise.resolve(new Date(firstDate)),
-            ]);
-
-            _.resetCache('subscriptions', 'united-states');
+            const cache = getCache(firstDate, 'united-states');
 
             return selectBannerTest(
                 Object.assign(targeting, {
@@ -76,18 +62,15 @@ describe('selectBannerTest', () => {
                 }),
                 tracking,
                 '',
+                getTests,
+                cache,
             ).then(result => {
                 expect(result).toBe(null);
             });
         });
 
         it('returns null if user is logged in and has a subscription', () => {
-            (cacheAsync as jest.Mock).mockReturnValue([
-                null,
-                (): Promise<Date> => Promise.resolve(new Date(firstDate)),
-            ]);
-
-            _.resetCache('subscriptions', 'united-states');
+            const cache = getCache(firstDate, 'united-states');
 
             return selectBannerTest(
                 Object.assign(targeting, {
@@ -95,6 +78,8 @@ describe('selectBannerTest', () => {
                 }),
                 tracking,
                 '',
+                getTests,
+                cache,
             ).then(result => {
                 expect(result).toBe(null);
             });
@@ -122,25 +107,15 @@ describe('selectBannerTest', () => {
             clientName: '',
         };
         it('returns banner if it has never been dismissed', () => {
-            (cacheAsync as jest.Mock).mockReturnValue([
-                null,
-                (): Promise<Date> => Promise.resolve(new Date(secondDate)),
-            ]);
+            const cache = getCache(secondDate, 'australia');
 
-            _.resetCache('subscriptions', 'australia');
-
-            return selectBannerTest(targeting, tracking, '').then(result => {
+            return selectBannerTest(targeting, tracking, '', getTests, cache).then(result => {
                 expect(result && result.test.name).toBe('GuardianWeeklyBanner');
             });
         });
 
         it('returns null if other contributions banner was dismissed and subs switch is off', () => {
-            (cacheAsync as jest.Mock).mockReturnValue([
-                null,
-                (): Promise<Date> => Promise.resolve(new Date(secondDate)),
-            ]);
-
-            _.resetCache('subscriptions', 'australia');
+            const cache = getCache(secondDate, 'australia');
 
             return selectBannerTest(
                 Object.assign(targeting, {
@@ -150,18 +125,15 @@ describe('selectBannerTest', () => {
                 }),
                 tracking,
                 '',
+                getTests,
+                cache,
             ).then(result => {
                 expect(result && result.test.name).toBe(null);
             });
         });
 
         it('returns banner if has been redeployed', () => {
-            (cacheAsync as jest.Mock).mockReturnValue([
-                null,
-                (): Promise<Date> => Promise.resolve(new Date(firstDate)),
-            ]);
-
-            _.resetCache('subscriptions', 'australia');
+            const cache = getCache(firstDate, 'australia');
 
             return selectBannerTest(
                 Object.assign(targeting, {
@@ -171,18 +143,15 @@ describe('selectBannerTest', () => {
                 }),
                 tracking,
                 '',
+                getTests,
+                cache,
             ).then(result => {
                 expect(result && result.test.name).toBe('GuardianWeeklyBanner');
             });
         });
 
         it('returns null if shouldHideReaderRevenue', () => {
-            (cacheAsync as jest.Mock).mockReturnValue([
-                null,
-                (): Promise<Date> => Promise.resolve(new Date(firstDate)),
-            ]);
-
-            _.resetCache('subscriptions', 'australia');
+            const cache = getCache(firstDate, 'australia');
 
             return selectBannerTest(
                 Object.assign(targeting, {
@@ -190,6 +159,8 @@ describe('selectBannerTest', () => {
                 }),
                 tracking,
                 '',
+                getTests,
+                cache,
             ).then(result => {
                 expect(result).toBe(null);
             });
